backend/ML-LLM-hybrid-recommendation-system/requirements.txt
pandas==2.3.0
numpy==2.3.1
scikit-learn==1.7.0
scipy==1.16.0
openai==1.93.0
python-dotenv==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
joblib==1.5.1
threadpoolctl==3.6.0
